Backend/Face_recog_backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
dlib==20.0.0
numpy==2.2.6
opencv-python==4.12.0.88
Pillow==11.3.0
requests==2.32.5
